Refactor Item for external input helpers
