+++ Assets/Scripts/Item.cs
@@ -2,22 +2,12 @@
 using UnityEngine;
 using UnityEditor;
 
-public enum SwipeDirection
-{
-    None,
-    Up,
-    Right,
-    Down,
-    Left
-}
-
 public class Item : MonoBehaviour
 {
     public int x; // Horizontal position
 
     public int y; // Vertical position
 
-    private Vector2 initialTouchPosition;
     private Vector2 finalTouchPosition;
 
     public int id;
@@ -80,62 +70,17 @@ public class Item : MonoBehaviour
         isFalling = false;
     }
 
-    void OnMouseDown()
-    {
-        float cameraZPos = Camera.main.transform.position.z;
-        initialTouchPosition = new Vector2(x, y);
-    }
-
     void OnMouseUp()
     {
+        // We get the camera to cancel the Z vector and ScreenToWorldPoint returns distances in the same Z value
         float cameraZPos = Camera.main.transform.position.z;
         finalTouchPosition = Camera.main.ScreenToWorldPoint(Input.mousePosition + new Vector3(0f, 0f, -cameraZPos));
 
-
-        float angle = CalculateAngle(initialTouchPosition, finalTouchPosition);
         if (OnMouseOverItemEventHandler != null)
         {
-            SwipeDirection direction = SwipeDirection.None;
-            float max = GetSwipeMajorAxis(initialTouchPosition, finalTouchPosition);
-            if (max > 0.5f)
-            {
-                if (angle > -45f && angle <= 45f)
-                {
-                    direction = SwipeDirection.Right;
-                }
-                else if (angle > 45f && angle <= 135f)
-                {
-                    direction = SwipeDirection.Up;
-                }
-                else if (angle > 135f || angle <= -135f)
-                {
-                    direction = SwipeDirection.Left;
-                }
-                else if (angle > -135f && angle <= -45f)
-                {
-                    direction = SwipeDirection.Down;
-                }
-            }
-
+            SwipeDirection direction = InputHelpers.GetSwipeDirection(new Vector2(x, y), finalTouchPosition, 0.5f);
             OnMouseOverItemEventHandler(this, direction);
         }
-    }
-
-    float CalculateAngle(Vector2 initial, Vector2 final)
-    {
-        float xDiff = final.x - initial.x;
-        float yDiff = final.y - initial.y;
-        float angle = Mathf.Atan2(yDiff, xDiff) * 180f / Mathf.PI;
-        Debug.Log("Found angle: " + angle);
-        return angle;
-    }
-
-    float GetSwipeMajorAxis(Vector2 initial, Vector2 final)
-    {
-        float maxX = Mathf.Abs(final.x - initial.x);
-        float maxY = Mathf.Abs(final.y - initial.y);
-
-        return Mathf.Max(maxX, maxY);
     }
 
 #if UNITY_EDITOR
