+++ Assets/Scripts/Tools/InputHelpers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc6b52b117f4d4a4ab3ed5228a604d4f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
